ajuste de permissao

+++ src/main/java/com/totem/service/UsuarioService.java
@@ -5,15 +5,19 @@ import java.util.List;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.totem.dto.UsuarioDTO;
 import com.totem.entity.Etapa;
 import com.totem.entity.Usuario;
+import com.totem.exception.CustomErrorException;
 import com.totem.repository.UsuarioRepository;
 
 @Service
 public class UsuarioService {
+	
+	private static final String ERRO_PERMISSAO = "Usuário sem permissão";  
 
 	@Autowired
 	private UsuarioRepository usuarioRepository;
@@ -21,7 +25,10 @@ public class UsuarioService {
 	@Autowired
 	private EtapaService etapaService;
 
-	public List<Usuario> listar() {
+	public List<Usuario> listar(String emailUsuario) {
+		if(!isAdm(emailUsuario)) {
+			throw new CustomErrorException(HttpStatus.UNAUTHORIZED, ERRO_PERMISSAO);
+		}
 		return usuarioRepository.findAll();
 	}
 
@@ -34,6 +41,10 @@ public class UsuarioService {
 	}
 
 	public Usuario salvar(UsuarioDTO usuarioDTO, String emailUsuario) {
+		
+		if(!isAdm(emailUsuario)) {
+			throw new CustomErrorException(HttpStatus.UNAUTHORIZED, ERRO_PERMISSAO);
+		}
 
 		Usuario usuario = new Usuario();
 
@@ -44,8 +55,20 @@ public class UsuarioService {
 		usuarioRepository.save(usuario);
 		return usuario;
 	}
+	
+	public Usuario salvar(UsuarioDTO usuarioDTO) {
+		
+		Usuario usuario = new Usuario();
+		BeanUtils.copyProperties(usuarioDTO, usuario);
+		usuarioRepository.save(usuario);
+		return usuario;
+	}
 
 	private void addListEtapaInUsuario(UsuarioDTO usuarioDTO, Usuario usuario) {
+		
+		if(!isAdm(usuario.getEmail())) {
+			throw new CustomErrorException(HttpStatus.UNAUTHORIZED, ERRO_PERMISSAO);
+		}
 		
 		if (usuarioDTO.getEtapaList() == null || usuarioDTO.getEtapaList().isEmpty()) {
 			return;
@@ -58,7 +81,11 @@ public class UsuarioService {
 		usuario.setEtapaList(etapaList);
 	}
 
-	public Usuario delete(Long id) {
+	public Usuario delete(Long id, String emailUsuario) {
+		
+		if(!isAdm(emailUsuario)) {
+			throw new CustomErrorException(HttpStatus.UNAUTHORIZED, ERRO_PERMISSAO);
+		}
 		Usuario usuario = usuarioRepository.findById(id).get();
 		usuarioRepository.delete(usuario);
 		return usuario;
@@ -69,7 +96,7 @@ public class UsuarioService {
 		if(usuario == null) {
 			return false;
 		}
-		return usuario.getIsAdmin()==null?false:usuario.getIsAdmin();
+		return usuario.getIsAdmin()==null?Boolean.FALSE:usuario.getIsAdmin();
 	}
 
 }
